backend/requirements.txt
fastapi
uvicorn[standard]
mongoengine
pydantic
requests
diffusers
torch
Pillow
gtts
moviepy
transformers
accelerate